requirements.txt
aiogram
aiogram_bot
alembic
apscheduler
database
python-dotenv
fastapi
llm
main
normalizer
parser
pathlib
pickle
psycopg2-binary
pydantic
requests
spacy
SQLAlchemy
telethon
telethon_client
utils
uvicorn
web